user is typing..
Homework: User is typing. Created a function request from client to server to get the username. Created input AddEventListener on keypress.. if they write anything or press a key inside the input.. the input value on userTyping changes to the username we got from the server. After 3 sec it goes back to null.